move Avatar component into new directory
