+++ packages/apollos-ui-kit/src/Avatar/Avatar/Avatar.tests.js
