+++ packages/apollos-ui-kit/src/index.js
@@ -5,7 +5,7 @@ export ActivityIndicator, {
   InlineActivityIndicator,
 } from './ActivityIndicator';
 export ActionBar, { ActionBarItem } from './ActionBar';
-export Avatar, { AvatarList } from './Avatar';
+export Avatar, { AvatarList } from './avatar';
 export BackgroundView from './BackgroundView';
 export Button, { ButtonLink, ButtonIcon } from './Button';
 export CampusCard from './CampusCard';
